requirements.txt
psycopg2-binary
restosaur==0.7.0b7
commonmark
django-sqltemplate==0.5.2
